replace float by double
git-svn-id: http://svn.cern.ch/guest/lhcb/Hlt/trunk@81486 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: 941970ee0ecdda4a3d2acfd2cc341e8790e9d2bf

+++ Hlt/Hlt/HltBeamGasAlley/src/BeamGasTrigClusterCut.cpp
@@ -79,7 +79,7 @@ StatusCode BeamGasTrigClusterCut::execute()
     }
   } 
   
-  float frUnusedRClusters = nRHits ? ( ( nRHits - nUsedRHits ) / (float)nRHits ) : 0.0001 ;
+  double frUnusedRClusters = nRHits ? ( ( nRHits - nUsedRHits ) / (double)nRHits ) : 0.0001 ;
    
   bool trigPass = false;
   if ( frUnusedRClusters > m_fracUnusedClustersCut ) trigPass = true;
